add github ci (second try)
