+++ .gitignore
@@ -1,3 +1,6 @@
 *.ipynb_checkpoints
 *__pycache__
 .vscode
+*.coverage
+*.pytest_cache
+ml_project.egg-info
