+++ README.md
@@ -1,3 +1,14 @@
 # Homework 1
 
  Использовался стандартный датасет: https://www.kaggle.com/datasets/cherngs heart-disease-cleveland-uci
+
+ ## run training code
+ 
+ `python src/main.py --train -d path/to/training/data.csv`
+
+ example: `./data/heart_cleveland_upload.csv`
+
+## run testing code
+`python src/main.py --inference -d path/to/inference/data.csv`
+
+example: `./data/testing.csv`
