+++ ml_project/requirements.txt
@@ -1,0 +1,5 @@
+numpy
+matplotlib
+seaborn
+sklearn
+pandas
